break down pages into smaller units
