+++ src/app/(secure)/layout.tsx
@@ -1,0 +1,7 @@
+export default function SecuredLayout({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  return <>{children}</>;
+}
